[FEAT #32] 실시간 검사 후 최종 결과 저장 API

+++ src/main/java/com/facemind/app/domain/Result.java
@@ -27,7 +27,7 @@ public class Result {
     private Integer heartRateAvg;
 
     @Column(nullable = false)
-    private Integer stressRate;
+    private Double stressRate;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id")
